+++ rx_notifier/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.2.0+2] - 2023-03-05
+* Added [rxNext] function for wait the next change of a [RxNotifer].
+* Added [RxNotifier.next];
+* Refactor: [RxAction].
+
 ## [2.1.0] - 2023-03-01
 * Added [RxCallback] Widget.
 * Added `context.callback()`.